fix(browser): satisfy dead-code boundaries
